backend/requirements.txt
fastapi==0.104.1              # Web API フレームワーク（メイン）
uvicorn==0.24.0               # サーバー（FastAPI を動かす）
sqlalchemy==2.0.23            # データベース操作（ORM）
python-multipart==0.0.6       # ファイルアップロード対応
python-jose[cryptography]==3.3.0  # JWT 認証
passlib[bcrypt]==1.7.4        # パスワード暗号化
python-decouple==3.8          # 環境変数管理
python-dotenv
